+++ 3-Tier-Architecture/modules/network/outputs.tf
@@ -1,36 +1,31 @@
-# Output for VPC id✅
+
 output "vpc_id" {
   value = aws_vpc.main.id
 }
 
-# Outputs for Public Subnets ✅
-output "public_subnet_az1_id" {
-  value = aws_subnet.public_subnet_az1.id
-}
-
-output "public_subnet_az2_id" {
-  value = aws_subnet.public_subnet_az2
-}
-
-# Output for Internet Gateway ✅
 output "internet_gateway" {
   value = aws_internet_gateway.main
 }
 
-# Outputs for Private Subnets ✅
-output "private_subnet_az1_id" {
-  value = aws_subnet.private_subnet_az1.id
+output "nat_gateway" {
+  value = aws_nat_gateway.main
+}
+output "public_subnet_id" {
+  value = aws_subnet.public.*.id
 }
 
-output "private_subnet_az2_id" {
-  value = aws_subnet.private_subnet_az2.id
+output "private_subnet_id" {
+  value = aws_subnet.private.*.id
 }
 
-# Outputs for Database Subnets ✅
-output "database_subnet_az1_id" {
-  value = aws_subnet.db_az1.id
+output "private_subnet_database_id" {
+  value = aws_subnet.db.*.id
 }
 
-output "database_subnet_az2_id" {
-  value = aws_subnet.db_az2.id
+output "rds_subnet_group_id" {
+  value = aws_db_subnet_group.rds.*.id
+}
+
+output "db_subnet_group_name" {
+  value = aws_db_subnet_group.rds.*.name
 }